requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
boto3==1.18.53
botocore==1.21.53
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.1
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.7
django-autocomplete-light==3.8.2
django-debug-toolbar==3.2.2
django-guardian==2.4.0
django-heroku==0.3.1
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-registration==3.2
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
jmespath==0.10.0
jsonfield==3.1.0
msgpack==1.0.2
Pillow==8.3.2
pinax-badges==3.0.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
swapper==1.1.2.post1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
zope.interface==5.4.0
django-admin-honeypot==1.1.0
django-environ==0.7.0